# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HEAVENx API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /users/{id}/reset-device:
    post:
      operationId: resetUserDevice
      tags: [users]
      summary: Reset device lock for user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device reset
  /users/progress:
    post:
      operationId: saveProgress
      tags: [users]
      summary: Save reading progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressInput"
      responses:
        "200":
          description: Progress saved
  /users/favorites/{seriesId}:
    post:
      operationId: addFavorite
      tags: [users]
      summary: Add series to favorites
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Added to favorites
    delete:
      operationId: removeFavorite
      tags: [users]
      summary: Remove series from favorites
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
...
  /comments/chapter/{chapterId}:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Post comment on chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete comment (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /social:
    get:
      operationId: listSocialLinks
      tags: [social]
      summary: List social media links
      responses:
        "200":
          description: Social links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialLink"
    post:
      operationId: createSocialLink
      tags: [social]
      summary: Create social link (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialLinkInput"
      responses:
        "201":
          description: Social link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialLink"
  /social/{id}:
    patch:
      operationId: updateSocialLink
      tags: [social]
      summary: Update social link (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialLinkUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialLink"
    delete:
      operationId: deleteSocialLink
      tags: [social]
      summary: Delete social link (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /popups:
    get:
      operationId: listPopups
      tags: [popup]
      summary: List popups
      responses:
        "200":
          description: Popups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Popup"
    post:
      operationId: createPopup
      tags: [popup]
      summary: Create popup (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PopupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Popup"
  /popups/{id}:
    patch:
      operationId: updatePopup
      tags: [popup]
      summary: Update popup (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PopupUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Popup"
    delete:
      operationId: deletePopup
      tags: [popup]
      summary: Delete popup (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    AdminUser:
      type: object
      required: [id, username, role, isActive]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        isActive:
          type: boolean
        language:
          type: string
        deviceLockMode:
          type: string
        deviceId:
          type: ["string", "null"]
        createdAt:
          type: string
        lastLoginAt:
          type: ["string", "null"]
    UserInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          type: string
        language:
          type: string
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        language:
          type: string
        deviceLockMode:
          type: string
...
    ProgressInput:
      type: object
      required: [seriesId, chapterId, page]
      properties:
        seriesId:
          type: string
        chapterId:
          type: string
        page:
          type: integer
    Comment:
      type: object
      required: [id, chapterId, userId, text, createdAt]
      properties:
        id:
          type: string
        chapterId:
          type: string
        userId:
          type: string
        username:
          type: string
        text:
          type: string
        isHidden:
          type: boolean
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
...
    SocialLink:
      type: object
      required: [id, platform, url, isEnabled]
      properties:
        id:
          type: string
        platform:
          type: string
        url:
          type: string
        label:
          type: ["string", "null"]
        isEnabled:
          type: boolean
        sortOrder:
          type: integer
    SocialLinkInput:
      type: object
      required: [platform, url]
      properties:
        platform:
          type: string
        url:
          type: string
        label:
          type: string
        isEnabled:
          type: boolean
        sortOrder:
          type: integer
    SocialLinkUpdate:
      type: object
      properties:
        platform:
          type: string
        url:
          type: string
        label:
          type: string
        isEnabled:
          type: boolean
        sortOrder:
          type: integer
    Popup:
      type: object
      required: [id, title, content, isEnabled]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        buttonText:
          type: ["string", "null"]
        buttonUrl:
          type: ["string", "null"]
        closeText:
          type: ["string", "null"]
        isEnabled:
          type: boolean
        scheduleStart:
          type: ["string", "null"]
        scheduleEnd:
          type: ["string", "null"]
        createdAt:
          type: string
    PopupInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        buttonText:
          type: string
        buttonUrl:
          type: string
        closeText:
          type: string
        isEnabled:
          type: boolean
        scheduleStart:
          type: string
        scheduleEnd:
          type: string
    PopupUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        buttonText:
          type: string
        buttonUrl:
          type: string
        closeText:
          type: string
        isEnabled:
          type: boolean
        scheduleStart:
          type: string
        scheduleEnd:
          type: string
...