# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * HEAVENx API
 * OpenAPI spec version: 0.1.0
 */

export interface UserInput {
  username: string;
  password: string;
  email?: string;
  role?: string;
  language?: string;
}