# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * HEAVENx API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  username?: string;
  password?: string;
  email?: string;
  role?: string;
  isActive?: boolean;
  language?: string;
  deviceLockMode?: string;
}